requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
blis==0.7.9
cachetools==5.3.0
catalogue==1.0.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
deepface==0.0.79
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
executing==1.2.0
filelock==3.11.0
fire==0.5.0
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gdown==4.7.1
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
imutils==0.5.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.0.4
mtcnn==0.1.1
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathy==0.10.1
pdfminer.six==20221105
pdfplumber==0.8.0
phonenumbers==8.13.8
pickleshare==0.7.5
Pillow==9.5.0
plac==1.1.3
platformdirs==3.2.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.14.0
pyparsing==3.0.9
pyresparser==1.0.6
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
resume-parser==0.8.4
retina-face==0.0.13
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==1.0.6
stack-data==0.6.2
stemming==1.0.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tesseract==0.1.3
thinc==7.4.6
tifffile==2023.3.21
tika==2.6.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Wand==0.6.11
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
